refactor: convert layout and home to Server Components (closes #14)
Move "use client" off the root layout, the home route, and the
ProjectPlaceholder leaf. Split /about and /projects route files into
thin server wrappers that render the existing client content, so the
route page exports are Server Components.

Root layout now uses an inline Script for gtag init instead of a
useEffect (functionally equivalent, the standard gtag.js snippet).

Bundle (Next 15 build output):
- /         page bundle 279 B -> 140 B
- /about    page bundle 177 B -> 139 B
- /projects page bundle 4.75 kB -> 4.51 kB

Pages that stay client (need state, events, or browser APIs):
contact form, image modal, navbar, language provider, project card,
playx1, banca-do-ingresso.
